remove data code, add json, make IndexedDocumentation more general
