+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = [ "analysis", "codespan", "data","lsp", "parser"]
+members = [ "analysis", "codespan", "lsp", "parser"]
 resolver = "2"
