--- data/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "data"
-version = "0.1.0"
-edition = "2024"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-quote = "1.0.40"
-proc-macro2 = "1.0.94"
