+++ data/ca65-keyword-doc.json
@@ -1,5 +1,5 @@
 {
-  "keywords_to_markdown": {
+  "keys_to_doc": {
     ".DEFINEDMACRO": "`.DEFINEDMACRO`\n---\nBuiltin function. The function expects an identifier as argument in parentheses.\nThe argument is evaluated, and the function yields \"true\" if the identifier\nalready has been defined as the name of a macro. Otherwise, the function yields\nfalse. Example:\n\n\n```ca65\n  .macro add foo\n          clc\n          adc foo\n  .endmacro\n\n  .if     .definedmacro(add)\n          add #$01\n  .else\n          clc\n          adc #$01\n  .endif\n  \n```\n\n\n\n",
     ".FATAL": "`.FATAL`\n---\nForce an assembly error and terminate assembly. The assembler will output an\nerror message preceded by \"User error\" and will terminate assembly\nimmediately.\n\nThis command may be used to check for initial conditions that must be\nset before assembling a source file.\n\nExample:\n\n\n```ca65\n  .if     foo = 1\n  ...\n  .elseif bar = 1\n  ...\n  .else\n  .fatal  \"Must define foo or bar!\"\n  .endif\n  \n```\n\n\n\nSee also: [`.ERROR`](https://cc65.github.io/doc/ca65.html#.ERROR),\n[`.OUT`](https://cc65.github.io/doc/ca65.html#.OUT),\n[`.WARNING`](https://cc65.github.io/doc/ca65.html#.WARNING)\n\n",
     ".IFNBLANK": "`.IFNBLANK`\n---\nConditional assembly: Check if there are any remaining tokens in this line,\nand evaluate to TRUE if this is the case, and to FALSE otherwise. If the\ncondition is not true, further lines are not assembled until an [`.ELSE`](https://cc65.github.io/doc/ca65.html#.ELSE), [`.ELSEIF`](https://cc65.github.io/doc/ca65.html#.ELSEIF) or\n[`.ENDIF`](https://cc65.github.io/doc/ca65.html#.ENDIF) directive.\n\nThis command is often used to check if a macro parameter was given.\nSince an empty macro parameter will evaluate to nothing, the condition\nwill evaluate to FALSE if an empty parameter was given.\n\nExample:\n\n\n```ca65\n  .macro     arg1, arg2\n             lda     #arg1\n  .ifnblank  arg2\n             lda     #arg2\n  .endif\n  .endmacro\n  \n```\n\n\n\nSee also: [`.BLANK`](https://cc65.github.io/doc/ca65.html#.BLANK)\n\n",
@@ -143,7 +143,7 @@
     ".MACRO": "`.MAC, .MACRO`\n---\nStart a classic macro definition. The command is followed by an identifier\n(the macro name) and optionally by a comma separated list of identifiers\nthat are macro parameters. A macro definition is terminated by [`.ENDMACRO`](https://cc65.github.io/doc/ca65.html#.ENDMACRO).\n\nExample:\n\n\n```ca65\n  .macro  ldax    arg             ; Define macro ldax\n          lda     arg\n          ldx     arg+1\n  \n```\n\n\n\nSee: [`.DELMACRO`](https://cc65.github.io/doc/ca65.html#.DELMACRO),\n[`.ENDMACRO`](https://cc65.github.io/doc/ca65.html#.ENDMACRO),\n[`.EXITMACRO`](https://cc65.github.io/doc/ca65.html#.EXITMACRO)\n\nSee also section \n[Macros](https://cc65.github.io/doc/ca65.html#macros).\n\n",
     ".CONDES": "`.CONDES`\n---\nExport a symbol and mark it in a special way. The linker is able to build\ntables of all such symbols. This may be used to automatically create a list\nof functions needed to initialize linked library modules.\n\nNote: The linker has a feature to build a table of marked routines, but it\nis your code that must call these routines, so just declaring a symbol with\n`.CONDES` does nothing by itself.\n\nAll symbols are exported as an absolute (16 bit) symbol. You don't need to\nuse an additional [`.EXPORT`](https://cc65.github.io/doc/ca65.html#.EXPORT) statement, this\nis implied by `.CONDES`.\n\n`.CONDES` is followed by the type, which may be `constructor`,\n`destructor` or a numeric value between 0 and 6 (where 0 is the same as\nspecifying `constructor` and 1 is equal to specifying `destructor`).\nThe [`.CONSTRUCTOR`](https://cc65.github.io/doc/ca65.html#.CONSTRUCTOR), [`.DESTRUCTOR`](https://cc65.github.io/doc/ca65.html#.DESTRUCTOR) and [`.INTERRUPTOR`](https://cc65.github.io/doc/ca65.html#.INTERRUPTOR) commands are actually shortcuts for `.CONDES`\nwith a type of `constructor` resp. `destructor` or `interruptor`.\n\nAfter the type, an optional priority may be specified. Higher numeric values\nmean higher priority. If no priority is given, the default priority of 7 is\nused. Be careful when assigning priorities to your own module constructors\nso they won't interfere with the ones in the cc65 library.\n\nExample:\n\n\n```ca65\n  .condes         ModuleInit, constructor\n  .condes         ModInit, 0, 16\n  \n```\n\n\n\nSee the [`.CONSTRUCTOR`](https://cc65.github.io/doc/ca65.html#.CONSTRUCTOR), [`.DESTRUCTOR`](https://cc65.github.io/doc/ca65.html#.DESTRUCTOR) and [`.INTERRUPTOR`](https://cc65.github.io/doc/ca65.html#.INTERRUPTOR) commands and the separate section \n[Module constructors/destructors](https://cc65.github.io/doc/ca65.html#condes) explaining the feature in more\ndetail.\n\n"
   },
-  "aliases": {
+  "keys_with_shared_doc": {
     ".BYT": ".BYTE",
     ".REFTO": ".REFERTO",
     ".FILEOPT": ".FOPT",
