+++ lsp/Cargo.toml
@@ -15,7 +15,6 @@ toml = "0.8.19"
 tempfile = "3.14.0"
 streaming-iterator = "0.1.9"
 parser = { path = "../parser" }
-data = { path = "../data" }
 analysis = { path = "../analysis" }
 codespan = { path = "../codespan", features = ["lsp"] }
 lazy_static = "1.5.0"
