+++ lsp/src/asm_server.rs
@@ -1,4 +1,3 @@
-use crate::ca65_doc::CA65_DOC;
 use crate::codespan::{FileId, Files, IndexError};
 use crate::completion::{BlockControlCompletionProvider, CompletionProvider, InstructionCompletionProvider, SymbolCompletionProvider};
 use crate::configuration::{load_project_configuration, Configuration};
@@ -7,7 +6,7 @@ use crate::error::file_error_to_lsp;
 use crate::symbol_cache::{
     symbol_cache_get, symbol_cache_insert, symbol_cache_reset, SymbolType,
 };
-use crate::OPCODE_DOCUMENTATION;
+use crate::documentation::{CA65_DOC, INSTRUCTION_DOC};
 use analysis::ScopeKind;
 use parser::ParseError;
 use std::collections::HashMap;
@@ -235,16 +234,16 @@ impl LanguageServer for Asm {
                 .get_word_at_position(params.text_document_position_params.position.into())
                 .map_err(file_error_to_lsp)?;
 
-            if let Some(documentation) = OPCODE_DOCUMENTATION
+            if let Some(documentation) = INSTRUCTION_DOC
                 .get()
                 .unwrap()
-                .get(&word.to_string().to_lowercase())
+                .get_doc_for_word(&word.to_lowercase())
             {
                 return Ok(Some(Hover {
                     range: None,
                     contents: HoverContents::Markup(MarkupContent {
                         kind: MarkupKind::Markdown,
-                        value: documentation.clone(),
+                        value: documentation,
                     }),
                 }));
             }
