+++ lsp/src/main.rs
@@ -5,18 +5,14 @@ mod diagnostics;
 mod instructions;
 mod logger;
 mod symbol_cache;
-mod ca65_doc;
 mod completion;
 mod definition;
 mod error;
 mod path;
+mod documentation;
 
 use asm_server::Asm;
 use tower_lsp_server::{LspService, Server};
-
-use data::include_documentation;
-
-include_documentation!();
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
@@ -25,9 +21,7 @@ async fn main() -> anyhow::Result<()> {
 
     symbol_cache::init_symbol_cache();
     instructions::init_instruction_map();
-    ca65_doc::parse_json_to_hashmap();
-
-    documentation_init();
+    documentation::init_documentation_map();
 
     let (service, socket) = LspService::new(|client| {
         Asm::new(client)
